Renamed VideoCardList to ResultsList / Added Test to ResultList
